pyproject.toml
[project]
name = "with-grok"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=2026.1.4",
    "google-generativeai>=0.8.6",
    "gradio>=6.6.0",
    "httpx>=0.28.1",
    "openai>=2.21.0",
    "pip-system-certs>=5.3",
    "python-dotenv>=1.2.1",
]
